Implement execution of projects
The projects within the zip folders will now be executed, with output
directed to a output.txt file in the created UserID folder. If input is
required for execution, an input file can be redirected to the applicable
class at execution time.

Instructions for configuration of the execution can be found in the README
file of the root directory.

An 'initial' solution of handling byte arrays as strings by stripping the
surrounding byte characters was switched with the decode method to return
the correctly formatted string.

The batch file handling execution of projects was modified to handle
classes without a specified file to redirect.