Modify printed statements
The program will indicate when it starts and completes through print
statements.

A print statement has been added to identify the currently processing
assignment.

Print statements used in testing have been removed.

The print statement for execution output has been commented; this output
will be saved in the output file, and is not relevant during processing
(errors are more important).